# backend/openapi.yaml
openapi: 3.1.0
info:
  title: THUIE API
  version: 1.0.0-draft
  description: >-
    REST + WebSocket contract for the THUIE campus alumni info-exchange
    platform. Design skeleton - mirrors BACKEND.md and the PostgreSQL schema in
    PROJECT.md 6. Server implementation deferred to the build phase.
  license: { name: Proprietary }
servers:
  - url: https://api.thuie.example.com/api/v1
    description: Production (placeholder)
  - url: http://localhost:3000/api/v1
    description: Local development

# OpenAPI has no native WebSocket binding, so the realtime surface (BACKEND.md §9)
# is documented here as a vendor extension. Socket.IO server, namespace
# `/realtime`, authenticated with the same access JWT supplied via handshake
# `auth.token`, an `Authorization: Bearer` header, or a `?token=` query.
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get: { tags: [Users], operationId: getMe, summary: Current user, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Users], operationId: updateMe, summary: Update profile fields, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/UserUpdate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Users], operationId: deleteMe, summary: Soft delete + PII scrub, responses: { '204': { description: deleted }, default: { $ref: '#/components/responses/Error' } } }
  /me/role-strategy:
    get: { tags: [Users], operationId: getMyRoleStrategy, summary: Effective permission flags, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/RoleStrategy' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/push-token:
    post: { tags: [Users], operationId: registerPushToken, summary: Register FCM push token, requestBody: { content: { application/json: { schema: { type: object, properties: { token: { type: string }, platform: { type: string } } } } } }, responses: { '204': { description: stored }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Users], operationId: unregisterPushToken, summary: Unregister a push token, parameters: [ { in: query, name: token, required: true, schema: { type: string } } ], responses: { '204': { description: removed }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Departments ----
...
  /infos:
    get: { tags: [Infos], operationId: listInfos, summary: Info feed (cursor), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: category, schema: { $ref: '#/components/schemas/InfoCategory' } }, { $ref: '#/components/parameters/Sort' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Infos], operationId: createInfo, summary: Submit info (-> moderation), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InfoCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPost' } } } }, '422': { description: keyword blocked, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
...
  /infos/{id}/report:
    post: { tags: [Infos], operationId: reportInfo, summary: Report an info post, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreate' } } } }, responses: { '202': { description: filed }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Alumni ----
...
  /posts:
    get: { tags: [Forum], operationId: listPosts, summary: Forum feed (cursor), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: tag, schema: { type: string } }, { in: query, name: q, schema: { type: string } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Forum], operationId: createPost, summary: Submit post (-> moderation), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/PostCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
...
  /mentors:
    get: { tags: [Mentorship], operationId: listMentors, summary: Browse mentors (cursor), parameters: [ { in: query, name: area, schema: { type: string } }, { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MentorProfile' } } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/mentor-profile:
    post: { tags: [Mentorship], operationId: createMyMentorProfile, summary: Become a mentor, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/MentorCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/MentorProfile' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /mentorship/applications:
    get: { tags: [Mentorship], operationId: listApplications, summary: List applications (incoming/outgoing, status filter, cursor), parameters: [ { in: query, name: role, schema: { type: string, enum: [incoming, outgoing] } }, { in: query, name: status, schema: { $ref: '#/components/schemas/MentorshipStatus' } }, { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MentorshipApplication' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Mentorship], operationId: createApplication, summary: Apply to a mentor, parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { mentor_user_id: { type: string }, message: { type: string } } } } } }, responses: { '201': { description: pending, content: { application/json: { schema: { $ref: '#/components/schemas/MentorshipApplication' } } } }, '409': { description: duplicate active pair, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /mentorship/applications/{id}:
    patch: { tags: [Mentorship], operationId: respondApplication, summary: Mentor accept / reject, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { status: { type: string, enum: [accepted, rejected] } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/MentorshipApplication' } } } }, '409': { description: mentor full, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Mentorship], operationId: withdrawApplication, summary: Mentee withdraw / end, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: ok }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Engagement ----
...
  /me/sessions:
    get: { tags: [Users], operationId: listSessions, summary: Active device sessions, responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Session' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Users], operationId: revokeAllSessions, summary: Revoke all other sessions, responses: { '204': { description: revoked }, default: { $ref: '#/components/responses/Error' } } }
  /me/sessions/{id}:
    delete: { tags: [Users], operationId: revokeSession, summary: Revoke one session, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: revoked }, default: { $ref: '#/components/responses/Error' } } }
  /me/infos:
    get: { tags: [Infos], operationId: listMyInfos, summary: Own infos (all statuses + history), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/posts:
    get: { tags: [Forum], operationId: listMyPosts, summary: Own posts (all statuses), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/comments:
    get: { tags: [Forum], operationId: listMyComments, summary: Own comments, parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/CommentPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
...
  /me/notification-preferences:
    get: { tags: [Notifications], operationId: getNotificationPrefs, summary: Per-type mute, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPreferences' } } } }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Notifications], operationId: updateNotificationPrefs, summary: Mute / unmute one notification type, requestBody: { content: { application/json: { schema: { type: object, required: [type, muted], properties: { type: { $ref: '#/components/schemas/NotificationType' }, muted: { type: boolean } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationPreferences' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Alumni visibility request ----
...
  /reports:
    post: { tags: [Engagement], operationId: createReport, summary: File a report (any target), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreateGeneric' } } } }, responses: { '202': { description: filed }, default: { $ref: '#/components/responses/Error' } } }
  /me/reports:
    get: { tags: [Engagement], operationId: listMyReports, summary: Filed reports + resolution, parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ReportPage' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Messaging group participants ----
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    Id: { in: path, name: id, required: true, schema: { type: string } }
...
    Cursor: { in: query, name: cursor, schema: { type: string }, description: Opaque keyset cursor }
    Sort: { in: query, name: sort, schema: { type: string }, description: "-created_at (leading - = DESC)" }
    IdempotencyKey: { in: header, name: Idempotency-Key, schema: { type: string, format: uuid } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    Role: { type: string, enum: [student, graduate, admin, admin_super] }
    UserStatus: { type: string, enum: [unverified, active, posting_restricted, banned, deleted] }
...
    ContentStatus: { type: string, enum: [pending, approved, rejected, taken_down] }
...
    Visibility: { type: string, enum: [student_only, all, admin_only] }
    InfoCategory: { type: string, enum: [internal, open, recruitment] }
    InfoSource: { type: string, enum: [official, user] }
...
    MentorshipStatus: { type: string, enum: [pending, accepted, rejected, withdrawn, ended] }
    ReportStatus: { type: string, enum: [open, resolved_ignored, resolved_deleted, resolved_restricted] }
    NotificationType: { type: string, enum: [review_result, comment_reply, connection, system, report_result, identity_change, broadcast] }
...
    ReportTargetType: { type: string, enum: [info_post, forum_post, comment, alumni_profile, user] }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine code }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
        errors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
    CursorMeta: { type: object, properties: { nextCursor: { type: [string, 'null'] }, limit: { type: integer } } }
    OffsetMeta: { type: object, properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer } } }
...
    Department: { type: object, properties: { id: { type: string }, code: { type: string }, name_zh: { type: string }, name_en: { type: string }, faculty: { type: [string, 'null'] } } }
    User: { type: object, properties: { id: { type: string }, name: { type: string }, role: { $ref: '#/components/schemas/Role' }, status: { $ref: '#/components/schemas/UserStatus' }, student_id: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, email: { type: [string, 'null'] }, department: { $ref: '#/components/schemas/Department' }, grade_year: { type: [string, 'null'] }, graduation_year: { type: [string, 'null'] }, avatar_url: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, wechat: { type: [string, 'null'], description: GET /me only }, whatsapp: { type: [string, 'null'], description: GET /me only }, linkedin: { type: [string, 'null'], description: GET /me only }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' }, is_online: { type: boolean }, created_at: { type: string, format: date-time } } }
    UserUpdate: { type: object, properties: { name: { type: string }, bio: { type: string }, avatar_media_id: { type: string }, grade_year: { type: string }, wechat: { type: string }, whatsapp: { type: string }, linkedin: { type: string }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' } } }
...
    RoleStrategy: { type: object, properties: { role: { $ref: '#/components/schemas/Role' }, can_view_info: { type: boolean }, can_view_internal: { type: boolean }, can_view_alumni: { type: boolean }, can_view_forum: { type: boolean }, can_submit_info: { type: boolean }, can_post_forum: { type: boolean }, can_comment: { type: boolean }, can_create_profile: { type: boolean } } }
    MediaObject: { type: object, properties: { id: { type: string }, kind: { type: string, enum: [image, attachment, avatar] }, url: { type: string }, file_name: { type: string }, mime_type: { type: string }, size_bytes: { type: integer }, width: { type: [integer, 'null'] }, height: { type: [integer, 'null'] }, thumbnail_key: { type: [string, 'null'] } } }
...
    InfoPost: { type: object, properties: { id: { type: string }, title: { type: string }, content: { type: string }, category: { $ref: '#/components/schemas/InfoCategory' }, source: { $ref: '#/components/schemas/InfoSource' }, visibility: { $ref: '#/components/schemas/Visibility' }, status: { $ref: '#/components/schemas/ContentStatus' }, pinned: { type: boolean }, version: { type: integer }, author: { $ref: '#/components/schemas/User' }, media: { type: array, items: { $ref: '#/components/schemas/MediaObject' } }, created_at: { type: string, format: date-time }, edited_at: { type: [string, 'null'], format: date-time } } }
    InfoCreate: { type: object, required: [title, content], properties: { title: { type: string, maxLength: 200 }, content: { type: string }, category: { $ref: '#/components/schemas/InfoCategory' }, visibility: { $ref: '#/components/schemas/Visibility' }, media_ids: { type: array, items: { type: string } } } }
...
    Tag: { type: object, properties: { id: { type: string }, name: { type: string }, slug: { type: string } } }
    ForumPost: { type: object, properties: { id: { type: string }, title: { type: string }, content: { type: string }, author: { $ref: '#/components/schemas/User' }, status: { $ref: '#/components/schemas/ContentStatus' }, location: { type: [string, 'null'] }, is_pinned: { type: boolean }, tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }, view_count: { type: integer }, like_count: { type: integer }, bookmark_count: { type: integer }, comment_count: { type: integer }, liked: { type: boolean }, bookmarked: { type: boolean }, liked_by: { type: array, maxItems: 3, items: { $ref: '#/components/schemas/User' } }, version: { type: integer }, created_at: { type: string, format: date-time }, last_comment_at: { type: [string, 'null'], format: date-time } } }
    PostCreate: { type: object, required: [title, content], properties: { title: { type: string }, content: { type: string }, location: { type: string }, tags: { type: array, items: { type: string } } } }
    Comment: { type: object, properties: { id: { type: string }, forum_post_id: { type: string }, parent_id: { type: [string, 'null'] }, author: { $ref: '#/components/schemas/User' }, content: { type: string }, status: { $ref: '#/components/schemas/ContentStatus' }, version: { type: integer }, created_at: { type: string, format: date-time } } }
...
    MentorProfile: { type: object, properties: { user: { $ref: '#/components/schemas/User' }, expertise: { type: string }, mentor_area: { type: string }, max_mentees: { type: integer }, current_mentees: { type: integer }, status: { type: string, enum: [pending, active, paused] } } }
    MentorCreate: { type: object, properties: { expertise: { type: string }, mentor_area: { type: string }, max_mentees: { type: integer, default: 3 } } }
    MentorshipApplication: { type: object, properties: { id: { type: string }, mentor: { $ref: '#/components/schemas/User' }, mentee: { $ref: '#/components/schemas/User' }, message: { type: string }, status: { $ref: '#/components/schemas/MentorshipStatus' }, responded_at: { type: [string, 'null'], format: date-time } } }
...
    Report: { type: object, properties: { id: { type: string }, reporter: { $ref: '#/components/schemas/User' }, target_type: { $ref: '#/components/schemas/ReportTargetType' }, target_id: { type: string }, reason: { type: string }, status: { $ref: '#/components/schemas/ReportStatus' }, result_note: { type: [string, 'null'] }, created_at: { type: string, format: date-time } } }
    ReportCreate: { type: object, required: [reason], properties: { reason: { type: string, maxLength: 500 } } }
...
    Session: { type: object, properties: { id: { type: string }, device_label: { type: [string, 'null'] }, ip: { type: [string, 'null'] }, current: { type: boolean }, created_at: { type: string, format: date-time }, last_used_at: { type: [string, 'null'], format: date-time } } }
...
    NotificationPreferences: { type: object, properties: { preferences: { type: array, items: { type: object, properties: { type: { $ref: '#/components/schemas/NotificationType' }, muted: { type: boolean } } } } } }
    ReportCreateGeneric: { type: object, required: [target_type, target_id, reason], properties: { target_type: { $ref: '#/components/schemas/ReportTargetType' }, target_id: { type: string }, reason: { type: string, maxLength: 500 } } }
...
    InfoPostPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/InfoPost' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...
    ForumPostPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/ForumPost' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
    CommentPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Comment' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...
    ReportPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Report' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
...